GOLDEN_AGE
This creates a bowling ball bouncing between the player and enemies. The bouncing is interesting, but collisions with enemies generate quite unpredictable behavior - where, for instance, the ball can be used as a shield to push enemies away, or is most efficient when "carried" by crowds of enemies inflicting damage on themselves.

Might make more sense to not run a collider on bouncing at all. We keep this commit in case we want to go back.

(Also note weapons.js has disabled shooting for the sake of testing bowling damage.)