shrines.js
Shrines spawn periodically, and trigger an effect while the player is within their radius.

This started way overpowered. I'm choosing the balance point as roughly: "the player spends half of the time in the aura". Compare it with beacons that way.

This may or may not feel balanced, but we have to take into account the constraints of staying in a specific area, within an arbitrary timeframe - whereas beacons give you optionality as to when, if not where.

Much playtesting shall be required.