Shrines changes: - GOD_HAMMER_S 2x cd, 1/2 interval - STORM_SHRINE 16 to 20s cd, 200 to 250ms interval - HEALING_SHRINE 24 to 32s
GOD_HAMMER feels weak as it is due to the 2s interval. We'll make the shrine appear twice less, but it should feel much more powerful to see 2-3 hammers fall in a row

HEALING_SHRINE feels overpowered with STORM_VENGEANCE, but even on its own borders on too strong compared to other healing patterns, even with the constraints tied to shrines. We're increasing the cooldown to balance it out

STORM_SHRINE is at least slightly too strong compared to the offensive shrines like BERSERK and ARCHER. Despite the lack of control tied to storms, it feels prudent to increase cd and decrease interval